Creating MainLegislationScreen style file
