+++ src/Styles/LegislationStyles/LawNumber12982Styles.js
@@ -1,6 +1,5 @@
 import { StyleSheet } from 'react-native';
 
-
 const styles = StyleSheet.create({
     title: {
       textAlign: 'center',
